chore(backend): Refactor the application into MVC structure
