+++ backend/routes/drugRoutes.js
@@ -1,0 +1,8 @@
+import express from 'express';
+import { searchDrugs } from '../controllers/drugController.js';
+
+const router = express.Router()
+
+router.get('/',searchDrugs);
+
+export default router;
